+++ src/EntityGraphQL/Schema/EntityQuerySchemaException.cs
@@ -4,8 +4,6 @@ namespace EntityGraphQL.Schema;
 
 public class EntityQuerySchemaException : Exception
 {
-    public EntityQuerySchemaException() { }
-
     public EntityQuerySchemaException(string message)
         : base(message) { }